requirements.txt
selenium==4.1.3
webdriver-manager==3.5.4
pyyaml==6.0
pytest==7.1.1
pytest-html==3.1.1
behave==1.2.6